Added comment and renamed to be correct
